+++ Controller/Frontend/SearchController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Opifer\CmsBundle\Controller\Front;
+namespace Opifer\CmsBundle\Controller\Frontend;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Response;
