+++ Resources/config/routing/routing.yml
@@ -44,7 +44,7 @@ opifer_redirect:
 
 opifer_cms_home:
     pattern:  /
-    defaults: { _controller: OpiferCmsBundle:Front/Page:home }
+    defaults: { _controller: OpiferCmsBundle:Frontend/Page:home }
     options: {sitemap: true}
 
 opifer_cms_auto_complete:
@@ -56,10 +56,6 @@ opifer_cms_auto_complete:
 opifer_cms_login:
     pattern:  /login
     defaults: { _controller: FOSUserBundle:Security:login }
-
-opifer_cms_exception:
-    resource: "@OpiferCmsBundle/Controller/Front/ExceptionController.php"
-    type:     annotation
 
 PrestaSitemapBundle:
     resource: "@PrestaSitemapBundle/Resources/config/routing.yml"
