+++ ValueProvider/FormValueProvider.php
@@ -5,14 +5,14 @@ namespace Opifer\CmsBundle\ValueProvider;
 use Opifer\EavBundle\ValueProvider\AbstractValueProvider;
 use Opifer\EavBundle\ValueProvider\ValueProviderInterface;
 use Symfony\Component\Form\FormBuilderInterface;
-use Doctrine\ORM\EntityRepository;
 
 class FormValueProvider extends AbstractValueProvider implements ValueProviderInterface
 {
+    /** @var string */
     protected $formClass;
 
     /**
-     * @param $formClass
+     * @param string $formClass
      */
     public function __construct($formClass)
     {
@@ -30,18 +30,7 @@ class FormValueProvider extends AbstractValueProvider implements ValueProviderIn
             'multiple'      => false,
             'class'         => $this->formClass,
             'property'      => 'name',
-            //'query_builder' => function (EntityRepository $er) {
-            //    return $er->createQueryBuilder('t')
-            //        ->where('t.objectClass = :objectClass')
-            //        ->setParameter('objectClass', 'Opifer\CmsBundle\Entity\Post');
-            //}
         ]);
-        //$builder->add('value', 'text', [
-        //    'label' => 'Success page',
-        //    'attr' => [
-        //        'placeholder' => 'The URL the form has to redirect to after success',
-        //    ]
-        //]);
     }
 
     /**
